# backend/docs/openapi.yaml
openapi: 3.0.0
paths:
  /api/health:
    get:
      operationId: HealthController_check
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: ok
      summary: Liveness probe
      tags:
        - Health
  /api/users:
    post:
      operationId: UserController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '409':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/user-already-exists
                  title:
                    type: string
                    example: User Already Exists
                  status:
                    type: number
                    example: 409
                  detail:
                    type: string
                    example: User already exists with email john.doe@example.com
                  email:
                    type: string
                    example: john.doe@example.com
      summary: Register a new user
      tags:
        - Users
  /api/users/me:
    get:
      operationId: UserController_profile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    example: 550e8400-e29b-41d4-a716-446655440000
                  email:
                    type: string
                    example: john.doe@example.com
                  firstName:
                    type: string
                    example: John
                  lastName:
                    type: string
                    example: Doe
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: about:blank
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: number
                    example: 401
        '404':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/entity-not-found
                  title:
                    type: string
                    example: Entity not found
                  status:
                    type: number
                    example: 404
                  detail:
                    type: string
                    example: >-
                      Entity not found with id
                      550e8400-e29b-41d4-a716-446655440000
                  entityId:
                    type: string
                    example: 550e8400-e29b-41d4-a716-446655440000
      security:
        - bearer: []
      summary: Get current user profile
      tags:
        - Users
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '401':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/invalid-credentials
                  title:
                    type: string
                    example: Invalid Credentials
                  status:
                    type: number
                    example: 401
                  detail:
                    type: string
                    example: Invalid credentials provided.
      summary: Login with email and password
      tags:
        - Auth
  /api/password-resets:
    post:
      operationId: PasswordResetController_request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetDto'
      responses:
        '201':
          description: Reset link sent
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '404':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/user-not-found
                  title:
                    type: string
                    example: User Not Found
                  status:
                    type: number
                    example: 404
                  detail:
                    type: string
                    example: No user found with email john.doe@example.com
                  email:
                    type: string
                    example: john.doe@example.com
      summary: Ask for a password reset link by email
      tags:
        - Password Resets
  /api/password-resets/{token}/password:
    put:
      operationId: PasswordResetController_reset
      parameters:
        - name: token
          required: true
          in: path
          description: The secret carried by the reset link
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '204':
          description: Password changed
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '404':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/password-reset-not-found
                  title:
                    type: string
                    example: Password Reset Not Found
                  status:
                    type: number
                    example: 404
                  detail:
                    type: string
                    example: No password reset matches the link that was used
        '410':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/password-reset-expired
                  title:
                    type: string
                    example: Password Reset Expired
                  status:
                    type: number
                    example: 410
                  detail:
                    type: string
                    example: Password reset link expired at 2026-01-01T11:00:00.000Z
                  expiredAt:
                    type: string
                    example: '2026-01-01T11:00:00.000Z'
      summary: Choose a new password using a reset link
      tags:
        - Password Resets
info:
  title: API
  description: ''
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 12
          example: password12345
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - email
        - password
        - firstName
        - lastName
    LoginUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    RequestPasswordResetDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        password:
          type: string
          minLength: 12
          example: password12345
      required:
        - password